Refactor instagram login button
