+++ app/components/InstagramLoginButton/index.js
@@ -1,10 +1,9 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { TouchableOpacity, Text, StyleSheet, Dimensions } from "react-native";
+import { TouchableOpacity, Text } from "react-native";
 import InstagramIcon from "../SVG/InstagramIcon";
 import { LinearGradient } from "expo";
-
-const { height, width } = Dimensions.get("window");
+import styles from "./instagramStyles";
 
 function InstagramLoginButton({ iconHeight, iconWidth, iconFill, onPress }) {
   return (
@@ -21,32 +20,6 @@ function InstagramLoginButton({ iconHeight, iconWidth, iconFill, onPress }) {
     </TouchableOpacity>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    width: width * 0.9,
-    height: height * 0.0949625187,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 1,
-    shadowRadius: 1,
-    elevation: 2
-  },
-  button: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    flexDirection: "row",
-    borderRadius: 4,
-    padding: 15
-  },
-  text: {
-    color: "#ffffff",
-    fontWeight: "bold",
-    marginLeft: 10,
-    fontSize: 22
-  }
-});
 
 InstagramLoginButton.defaultProps = {
   iconFill: "#ffffff",
